tutorials/cloud-data-pipeline/DatabaseIngestor/requirements.txt
msgpack==1.0.0
kafka-python==2.0.1
pymongo==3.10.1 